+++ Services/NerdShopping.API/Migrations/SQLContextModelSnapshot.cs
@@ -22,7 +22,7 @@ namespace NerdShopping.API.Migrations
                 {
                     b.Property<long>("Id")
                         .ValueGeneratedOnAdd()
-                        .HasColumnType("bigint")
+                        .HasColumnType("int")
                         .HasColumnName("id")
                         .HasAnnotation("SqlServer:ValueGenerationStrategy", SqlServerValueGenerationStrategy.IdentityColumn);
